+++ canvas_sdk/effects/compound_medications/compound_medication.py
@@ -5,11 +5,10 @@ from pydantic_core import InitErrorDetails, PydanticCustomError
 
 from canvas_generated.messages.effects_pb2 import Effect
 from canvas_sdk.base import TrackableFieldsModel
-from canvas_sdk.effects.base import _AsyncEffectMixin
 from canvas_sdk.v1.data.compound_medication import CompoundMedication as CompoundMedicationModel
 
 
-class CompoundMedication(TrackableFieldsModel, _AsyncEffectMixin):
+class CompoundMedication(TrackableFieldsModel):
     """Effect to create or update a Compound Medication record."""
 
     class Meta:
