+++ canvas_sdk/effects/patient_facility_address/patient_facility_address.py
@@ -8,7 +8,6 @@ from pydantic_core import InitErrorDetails
 
 from canvas_generated.messages.effects_pb2 import Effect
 from canvas_sdk.base import TrackableFieldsModel
-from canvas_sdk.effects.base import _AsyncEffectMixin
 from canvas_sdk.v1.data.facility import Facility
 from canvas_sdk.v1.data.patient import Patient
 from canvas_sdk.v1.data.patient import PatientFacilityAddress as PatientFacilityAddressModel
@@ -21,7 +20,7 @@ class AddressType(StrEnum):
     BOTH = "both"
 
 
-class PatientFacilityAddress(TrackableFieldsModel, _AsyncEffectMixin):
+class PatientFacilityAddress(TrackableFieldsModel):
     """
     Effect to create, update, or delete a Patient Facility Address in Canvas.
